rename alert_types to rule_types
